requirements.txt
playwright
python-dotenv
notify-py
tkinter